requirements.txt
Django==1.9.1
sorl-thumbnail
django-tinymce
django-tinymce-filebrowser
Pillow
pytils
psycopg2
requests
requests_oauthlib
rollbar